+++ modules/dns-stalwart/src/providers/desec.rs
@@ -1,0 +1,31 @@
+use std::{collections::HashMap, sync::Arc};
+
+use dns_update::DnsUpdater;
+use ferron_core::providers::Provider;
+use ferron_dns::DnsContext;
+
+use crate::client::DnsStalwartClient;
+
+pub struct DesecDnsProvider;
+
+impl Provider<DnsContext<'static>> for DesecDnsProvider {
+    fn name(&self) -> &'static str {
+        "desec"
+    }
+
+    fn execute(&self, ctx: &mut DnsContext) -> Result<(), Box<dyn std::error::Error>> {
+        let api_token = ctx
+            .config
+            .get_value("api_token")
+            .and_then(|v| v.as_string_with_interpolations(&HashMap::new()))
+            .ok_or(anyhow::anyhow!(
+                "Missing or invalid API token for 'desec' DNS provider"
+            ))?;
+
+        ctx.client = Some(Arc::new(DnsStalwartClient::new(
+            DnsUpdater::new_desec(&api_token, None)?,
+            3600,
+        )));
+        Ok(())
+    }
+}