+++ modules/dns-stalwart/src/providers/spaceship.rs
@@ -1,0 +1,38 @@
+use std::{collections::HashMap, sync::Arc};
+
+use dns_update::DnsUpdater;
+use ferron_core::providers::Provider;
+use ferron_dns::DnsContext;
+
+use crate::client::DnsStalwartClient;
+
+pub struct SpaceshipDnsProvider;
+
+impl Provider<DnsContext<'static>> for SpaceshipDnsProvider {
+    fn name(&self) -> &'static str {
+        "spaceship"
+    }
+
+    fn execute(&self, ctx: &mut DnsContext) -> Result<(), Box<dyn std::error::Error>> {
+        let api_key = ctx
+            .config
+            .get_value("api_key")
+            .and_then(|v| v.as_string_with_interpolations(&HashMap::new()))
+            .ok_or(anyhow::anyhow!(
+                "Missing or invalid API key for 'spaceship' DNS provider"
+            ))?;
+        let api_secret = ctx
+            .config
+            .get_value("api_secret")
+            .and_then(|v| v.as_string_with_interpolations(&HashMap::new()))
+            .ok_or(anyhow::anyhow!(
+                "Missing or invalid API secret for 'spaceship' DNS provider"
+            ))?;
+
+        ctx.client = Some(Arc::new(DnsStalwartClient::new(
+            DnsUpdater::new_spaceship(&api_key, &api_secret, None)?,
+            1200,
+        )));
+        Ok(())
+    }
+}