+++ types/dns/src/lib.rs
@@ -102,7 +102,11 @@ pub trait DnsClient: Send + Sync {
     async fn update_record(&self, record: &DnsRecord) -> Result<(), DnsProviderError>;
 
     /// Deletes all records matching the given name and type.
-    async fn delete_record(&self, name: &str, record_type: &str) -> Result<(), DnsProviderError>;
+    async fn delete_record(
+        &self,
+        name: &str,
+        record_type: DnsRecordType,
+    ) -> Result<(), DnsProviderError>;
 }
 
 /// Context passed to DNS [`Provider`](ferron_core::providers::Provider) implementations.
@@ -131,4 +135,34 @@ pub struct DnsContext<'a> {
     pub config: &'a ServerConfigurationBlock,
     /// The initialized DNS client, set by the provider during [`execute`](ferron_core::providers::Provider::execute).
     pub client: Option<Arc<dyn DnsClient>>,
+}
+
+/// Separates subdomain from domain name.
+pub async fn separate_subdomain_from_domain_name(domain_name: &str) -> (String, String) {
+    let parts: Vec<&str> = domain_name
+        .strip_suffix(".")
+        .unwrap_or(domain_name)
+        .split('.')
+        .collect();
+    let resolver = hickory_resolver::Resolver::builder_tokio()
+        .unwrap_or(hickory_resolver::Resolver::builder_with_config(
+            hickory_resolver::config::ResolverConfig::default(),
+            hickory_resolver::name_server::TokioConnectionProvider::default(),
+        ))
+        .build();
+
+    for parts_index in 0..parts.len() {
+        if resolver
+            .soa_lookup(format!("{}.", parts[parts_index..].join(".")))
+            .await
+            .is_ok()
+        {
+            // SOA record found
+            let subdomain = parts[..parts_index].join(".");
+            let domain = parts[parts_index..].join(".");
+            return (subdomain, domain);
+        }
+    }
+
+    ("".to_string(), parts.join("."))
 }